#96 [mod] getHomyColorRes() 확장함수 util로 이동
